setup vitest for api
